Update to the latest CSWin32
Take the latest drop of CSWin32.

There were a number of changes, but the key ones were:

- VARIANT moved to another namespace
- CsWin32 wrapper methods that take strings no longer have unusued/reserved parameters
- HGLOBAL is now used in signatures

A few other things in this change:

- Added start-vs.cmd to the solution folder
- Pinned created Arrays in VARIANT to avoid a potential memory corruption
- Removed VARIANT IRecord related code that was ultimately going to throw

+++ src/System.Windows.Forms/src/System/Windows/Forms/Control.MetafileDCWrapper.cs
@@ -131,7 +131,7 @@ public partial class Control
                         biHeight = bmp.bmHeight,
                         biPlanes = 1,
                         biBitCount = bmp.bmBitsPixel,
-                        biCompression = BI_COMPRESSION.BI_RGB
+                        biCompression = (uint)BI_COMPRESSION.BI_RGB
                     };
 
                     // Include the palette for 256 color bitmaps
